Try a small cleanup, hiera_config_value => *rest
The hiera_config_value is used in *args (only) as
a the cache key, however as the array is passed around
(without ever unpacking it), it has to be passed through
to the un-cached version.

This change makes it the last argument, and so allows
the uncached version's method signature to be more clear.